Add successfull test for sending a new score
