+++ all.test.js
@@ -1,5 +1,13 @@
+const {showScoreList, addScoreAndSend} = require('./src/leader-board.js')
 
 
-// test('capitalize string successfully', () => {
-//     expect(capitalize("name")).toBe("Name");
-// });
+// console.log(showScoreList())
+// console.log(addScoreAndSend(250))
+
+test('Check that new Score is successfully saved', () => {
+    myValue = ''
+    addScoreAndSend(250).then(data => {
+        expect(data.result).toBe("Leaderboard score created correctly.");
+    })
+    
+});
