+++ src/index.js
@@ -14,6 +14,8 @@ import coinSound from './assets/coin.mp3'
 import jumpSound from './assets/jump.mp3'
 import loseSound from './assets/lose.mp3'
 
+import {addScoreAndSend} from './leader-board.js'
+
 
 var player;
 var stars;
@@ -198,6 +200,9 @@ function hitBomb(player, bomb) {
 
     player.anims.play('turn');
 
+    const aPromise = addScoreAndSend(score)
+    aPromise.then( data => console.log(data.result))
+
     gameOver = true;
 }
 
@@ -219,38 +224,8 @@ const config = {
     }
 };
 
-const showScoreList = () => {
-    allScores = fetch('https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/efrnqEJRF6YCD6jcqp0F/scores/')
-        .then(response => response.json())
-        .then(data => data.result)
-        .catch(err => {
-            console.error(err)
-        })
-    
-    return allScores
-}
 
-const addScoreAndSend = () => {
-
-    fetch('https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/efrnqEJRF6YCD6jcqp0F/scores/', {
-        method: 'POST',
-        mode: 'cors',
-        headers: {
-            'Content-Type': 'application/json'
-        },
-        body: JSON.stringify({
-            "user": "Another one",
-            "score": score
-        })
-    })
-        .then(response => response.json())
-        .then(data => {
-            console.log(data)
-        })
-        .catch(err => {
-            console.error(err)
-        })
-
-}
 
 const game = new Phaser.Game(config);
+
+
